A major refactor of Crafthead
This refactor mostly sets the stage for new functionality and improves the reliability of Crafthead. Chief among these is switching to Workers KV for caching profiles (instead of caching skins at the local Cloudflare data center).

This also cleans up the code a great deal, which is a nice plus.